Share IndexCache base-path checks
