+++ test/manual/dsv4/eval_dsv4_indexcache_suite.py
@@ -6,11 +6,18 @@ import argparse
 import json
 import shutil
 import subprocess
+import sys
 import time
 from dataclasses import dataclass
 from pathlib import Path
 
-import requests
+sys.path.insert(0, str(Path(__file__).parent))
+from indexcache_base_path import (
+    fetch_server_info,
+)
+from indexcache_base_path import (
+    validate_server_info_for_base_path as validate_base_server_info,
+)
 
 FORBIDDEN_TASKS = {"mmlu", "gsm8k"}
 
@@ -141,57 +148,12 @@ def build_eval_cmd(task: EvalTask, base_url: str, args) -> list[str]:
     raise ValueError(f"unknown runner {task.runner}")
 
 
-def _as_int(value) -> int | None:
-    if value is None or isinstance(value, bool):
-        return None
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def speculative_config_paths(value, path: str = "") -> list[str]:
-    if isinstance(value, dict):
-        paths = []
-        for key, item in value.items():
-            child_path = f"{path}.{key}" if path else str(key)
-            if key == "speculative_algorithm" and item:
-                paths.append(f"{child_path}={item}")
-            elif key in {
-                "speculative_num_steps",
-                "speculative_eagle_topk",
-                "speculative_num_draft_tokens",
-            }:
-                int_value = _as_int(item)
-                if int_value is not None and int_value > 0:
-                    paths.append(f"{child_path}={item}")
-            elif key == "enable_multi_layer_eagle" and item:
-                paths.append(f"{child_path}={item}")
-            paths.extend(speculative_config_paths(item, child_path))
-        return paths
-    if isinstance(value, list):
-        paths = []
-        for i, item in enumerate(value):
-            child_path = f"{path}[{i}]" if path else f"[{i}]"
-            paths.extend(speculative_config_paths(item, child_path))
-        return paths
-    return []
-
-
-def validate_server_info_for_base_path(server_info: dict, label: str) -> None:
-    speculative_paths = speculative_config_paths(server_info)
-    if speculative_paths:
-        raise RuntimeError(
-            f"quality eval endpoint {label!r} reports speculative decoding enabled; "
-            "disable EAGLE/spec decode for base-path IndexCache validation: "
-            + ", ".join(speculative_paths)
-        )
+def validate_server_info_for_base_path(server_info: dict, label: str) -> list[str]:
+    return validate_base_server_info(server_info, f"quality eval endpoint {label!r}")
 
 
 def validate_endpoint_for_base_path(label: str, base_url: str, timeout: int) -> dict:
-    response = requests.get(base_url.rstrip("/") + "/server_info", timeout=timeout)
-    response.raise_for_status()
-    server_info = response.json()
+    server_info = fetch_server_info(base_url, timeout)
     validate_server_info_for_base_path(server_info, label)
     return server_info
