+++ test/manual/dsv4/profile_dsv4_indexcache_endpoint.py
@@ -6,10 +6,17 @@ import argparse
 import importlib.util
 import json
 import random
+import sys
 import time
 from pathlib import Path
 
 import requests
+
+sys.path.insert(0, str(Path(__file__).parent))
+from indexcache_base_path import (
+    fetch_server_info,
+    validate_server_info_for_base_path,
+)
 
 
 def _load_profile_analyzer():
@@ -56,60 +63,6 @@ def post_json(
     base_url: str, path: str, payload: dict, timeout: int
 ) -> requests.Response:
     return requests.post(base_url.rstrip("/") + path, json=payload, timeout=timeout)
-
-
-def get_server_info(base_url: str, timeout: int) -> dict:
-    response = requests.get(base_url.rstrip("/") + "/server_info", timeout=timeout)
-    response.raise_for_status()
-    return response.json()
-
-
-def _as_int(value) -> int | None:
-    if value is None or isinstance(value, bool):
-        return None
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def _speculative_config_paths(value, path: str = "") -> list[str]:
-    if isinstance(value, dict):
-        paths = []
-        for key, item in value.items():
-            child_path = f"{path}.{key}" if path else str(key)
-            if key == "speculative_algorithm" and item:
-                paths.append(f"{child_path}={item}")
-            elif key in {
-                "speculative_num_steps",
-                "speculative_eagle_topk",
-                "speculative_num_draft_tokens",
-            }:
-                int_value = _as_int(item)
-                if int_value is not None and int_value > 0:
-                    paths.append(f"{child_path}={item}")
-            elif key == "enable_multi_layer_eagle" and item:
-                paths.append(f"{child_path}={item}")
-            paths.extend(_speculative_config_paths(item, child_path))
-        return paths
-    if isinstance(value, list):
-        paths = []
-        for i, item in enumerate(value):
-            child_path = f"{path}[{i}]" if path else f"[{i}]"
-            paths.extend(_speculative_config_paths(item, child_path))
-        return paths
-    return []
-
-
-def validate_server_info_for_base_path(server_info: dict) -> list[str]:
-    speculative_paths = _speculative_config_paths(server_info)
-    if speculative_paths:
-        raise RuntimeError(
-            "server /server_info reports speculative decoding enabled; "
-            "rerun base-path IndexCache validation with EAGLE/spec decode off: "
-            + ", ".join(speculative_paths)
-        )
-    return speculative_paths
 
 
 def send_generate(base_url: str, prompt: str, max_tokens: int, timeout: int) -> dict:
@@ -274,7 +227,7 @@ def main() -> None:
             "speculative_decode": "dry run; /server_info not queried",
         }
     else:
-        server_info = get_server_info(args.endpoint, args.timeout)
+        server_info = fetch_server_info(args.endpoint, args.timeout)
         validate_server_info_for_base_path(server_info)
         server_checks = {
             "server_info_checked": True,
